feat: share hooks through the shared package
