+++ packages/neuron/src/components/Footer/index.tsx
@@ -2,10 +2,10 @@ import { ComponentProps, FC } from 'react'
 import clsx from 'clsx'
 import { useTranslation } from 'react-i18next'
 import Link from 'next/link'
+import { useIsMobile } from '@magickbase-website/shared'
 import styles from './index.module.scss'
 import IconFullLogo from './full-logo.svg'
 import { Contacts } from '../Contacts'
-import { useIsMobile } from '../../hooks'
 import { LinkWithEffect } from '../UpsideDownEffect'
 import { api } from '../../utils'
