+++ packages/neuron/src/components/UpsideDownEffect/index.tsx
@@ -1,8 +1,8 @@
 import { ComponentProps, FC, ReactNode } from 'react'
 import Link from 'next/link'
 import clsx from 'clsx'
+import { useIsMobile } from '@magickbase-website/shared'
 import styles from './index.module.scss'
-import { useIsMobile } from '../../hooks'
 
 interface UpsideDownEffectProps {
   content?: ReactNode | null
