+++ packages/neuron/src/pages/changelog/[[...slug]].page.tsx
@@ -11,10 +11,10 @@ import { Release, getReleases } from '../../utils'
 import { Page } from '../../components/Page'
 import styles from './index.module.scss'
 import ImgNeuronLogo from './neuron-logo.png'
-import { LinkComponentProps, useMarkdownProps } from '../../hooks'
 import IconTop from './top.svg'
 import IconMore from './more.svg'
 import { LinkWithEffect } from '../../components/UpsideDownEffect'
+import { LinkComponentProps, useMarkdownProps } from '../../hooks'
 
 const RELEASES_PER_PAGE = 2
