+++ packages/neuron/src/pages/posts/HelpDocHeader/index.tsx
@@ -5,6 +5,7 @@ import { useObservableState } from 'observable-hooks'
 import { DocSearch } from '@docsearch/react'
 import '@docsearch/css'
 import { useTranslation } from 'react-i18next'
+import { useIsHydrated, useIsMobile } from '@magickbase-website/shared'
 import { appSettings } from '../../../services/AppSettings'
 import ImgNeuronLogo from './neuron-logo.png'
 import IconDaylight from './daylight.svg'
@@ -13,7 +14,6 @@ import IconMenu from './menu.svg'
 import IconClose from './close.svg'
 import styles from './index.module.scss'
 import { APPID, Post, SEARCH_KEY, TopLevelMenu, languages, removeURLOrigin } from '../../../utils'
-import { useIsHydrated, useIsMobile } from '../../../hooks'
 import { ClassifiedPosts } from '../ClassifiedPosts'
 import { LanguageMenu } from './LanguageMenu'
 import { LanguageList } from './LanguageList'
