+++ packages/shared/src/hooks/useHighPrecisionScrollbarWidth.tsx
