+++ packages/shared/src/styles/variables.module.scss
@@ -1,4 +1,3 @@
-// TODO: The code in the shared module should be reused.
 $mobileBreakPoint: 750px;
 
 :export {
